requirements.txt
asgiref==3.5.2
astroid==2.11.7
autopep8==1.7.0
certifi==2022.6.15
charset-normalizer==2.1.1
colorama==0.4.5
dill==0.3.5.1
dj-database-url==1.0.0
Django==4.1
django-bootstrap-modal-forms==2.2.0
django-embed-video==1.4.4
django-on-heroku==1.1.2
django-widget-tweaks==1.4.12
gunicorn==20.1.0
idna==3.3
isort==5.10.1
lazy-object-proxy==1.7.1
mccabe==0.7.0
Pillow==9.2.0
platformdirs==2.5.2
psycopg2-binary==2.9.3
pycodestyle==2.9.1
pylint==2.14.5
python-decouple==3.6
requests==2.28.1
sqlparse==0.4.2
toml==0.10.2
tomli==2.0.1
tomlkit==0.11.4
tzdata==2022.2
urllib3==1.26.12
whitenoise==6.2.0
wrapt==1.14.1
